# Improve environment handling performance and cleanup API