--- a/Jint/Runtime/Interpreter/Expressions/JintAssignmentExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintAssignmentExpression.cs
@@ -48,7 +48,6 @@ namespace Jint.Runtime.Interpreter.Expressions
             if (_leftIdentifier is not null && JintEnvironment.TryGetIdentifierEnvironmentWithBindingValue(
                     engine.ExecutionContext.LexicalEnvironment,
                     _leftIdentifier.Identifier,
-                    StrictModeScope.IsStrictModeCode,
                     out var identifierEnvironment,
                     out var temp))
             {