--- a/Jint/Runtime/Interpreter/Expressions/JintUpdateExpression.cs
+++ b/Jint/Runtime/Interpreter/Expressions/JintUpdateExpression.cs
@@ -1,5 +1,6 @@
 using Jint.Native;
 using Jint.Runtime.Environments;
+using Environment = Jint.Runtime.Environments.Environment;
 
 namespace Jint.Runtime.Interpreter.Expressions
 {
@@ -121,20 +122,16 @@ namespace Jint.Runtime.Interpreter.Expressions
 
         private JsValue? UpdateIdentifier(EvaluationContext context)
         {
-            var strict = StrictModeScope.IsStrictModeCode;
             var name = _leftIdentifier!.Identifier;
-            var engine = context.Engine;
-            var env = engine.ExecutionContext.LexicalEnvironment;
             if (JintEnvironment.TryGetIdentifierEnvironmentWithBindingValue(
-                env,
-                name,
-                strict,
-                out var environmentRecord,
-                out var value))
+                    context.Engine.ExecutionContext.LexicalEnvironment,
+                    name,
+                    out var environmentRecord,
+                    out var value))
             {
-                if (strict && _evalOrArguments)
+                if (_evalOrArguments && StrictModeScope.IsStrictModeCode)
                 {
-                    ExceptionHelper.ThrowSyntaxError(engine.Realm);
+                    ExceptionHelper.ThrowSyntaxError(context.Engine.Realm);
                 }
 
                 var isInteger = value._type == InternalTypes.Integer;
@@ -167,7 +164,7 @@ namespace Jint.Runtime.Interpreter.Expressions
                     }
                 }
 
-                environmentRecord.SetMutableBinding(name.Key.Name, newValue!, strict);
+                environmentRecord.SetMutableBinding(name.Key, newValue!, StrictModeScope.IsStrictModeCode);
                 if (_prefix)
                 {
                     return newValue;