## Jint.Benchmark.DromaeoBenchmark

| **Diff**|Method|Prepared|Toolchain|FileName|Mean|Error|Allocated|
|------- |-------|-------|-------|-------|-------:|-------|-------:|
| Old |Run|False|Default|dromaeo-3d-cube|22.222 ms|0.2079 ms|6304.91 KB|
| **New** |	| **False** | **Default** |	| **19.953 ms (-10%)** | **0.2049 ms** | **6304.9 KB (0%)** |
| Old |Run|False|Default|dromaeo-core-eval|4.296 ms|0.0265 ms|327.85 KB|
| **New** |	| **False** | **Default** |	| **3.624 ms (-16%)** | **0.0306 ms** | **327.84 KB (0%)** |
| Old |Run|False|Default|dromaeo-object-array|42.425 ms|0.1915 ms|100367.72 KB|
| **New** |	| **False** | **Default** |	| **41.721 ms (-2%)** | **0.3537 ms** | **100367.76 KB (0%)** |
| Old |Run|False|Default|droma(...)egexp [21]|149.899 ms|1.0581 ms|151325.94 KB|
| **New** |	| **False** | **Default** |	| **147.555 ms (-2%)** | **1.1606 ms** | **151475.38 KB (0%)** |
| Old |Run|False|Default|droma(...)tring [21]|255.347 ms|5.2541 ms|1316020.46 KB|
| **New** |	| **False** | **Default** |	| **238.086 ms (-7%)** | **4.7367 ms** | **1315806.83 KB (0%)** |
| Old |Run|False|Default|droma(...)ase64 [21]|49.009 ms|0.3184 ms|2367.98 KB|
| **New** |	| **False** | **Default** |	| **45.742 ms (-7%)** | **0.1953 ms** | **2367.95 KB (0%)** |
| Old |Run|True|Default|dromaeo-3d-cube|21.630 ms|0.0634 ms|5966.65 KB|
| **New** |	| **True** | **Default** |	| **18.718 ms (-13%)** | **0.1213 ms** | **5966.66 KB (0%)** |
| Old |Run|True|Default|dromaeo-core-eval|4.554 ms|0.0182 ms|312.77 KB|
| **New** |	| **True** | **Default** |	| **5.395 ms (+18%)** | **0.0273 ms** | **312.78 KB (0%)** |
| Old |Run|True|Default|dromaeo-object-array|45.505 ms|0.2457 ms|100320.31 KB|
| **New** |	| **True** | **Default** |	| **41.496 ms (-9%)** | **0.4877 ms** | **100320.36 KB (0%)** |
| Old |Run|True|Default|droma(...)egexp [21]|110.251 ms|2.1159 ms|148277.7 KB|
| **New** |	| **True** | **Default** |	| **109.215 ms (-1%)** | **0.9305 ms** | **152477.6 KB (+3%)** |
| Old |Run|True|Default|droma(...)tring [21]|252.442 ms|5.3163 ms|1315823.89 KB|
| **New** |	| **True** | **Default** |	| **239.481 ms (-5%)** | **5.0117 ms** | **1315753.4 KB (0%)** |
| Old |Run|True|Default|droma(...)ase64 [21]|45.947 ms|0.2130 ms|2268.82 KB|
| **New** |	| **True** | **Default** |	| **42.438 ms (-8%)** | **0.2702 ms** | **2268.79 KB (0%)** |